% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_frame_in_pkg.R
\name{check_data_frame_in_pkg}
\alias{check_data_frame_in_pkg}
\title{Check if package contains data.frames}
\usage{
check_data_frame_in_pkg(pkg)
}
\arguments{
\item{pkg}{name of package (a character vector)}
}
\value{
logical (\code{TRUE} = has data.frame, \code{FALSE} = no data.frame)
}
\description{
Returns \code{TRUE} if package has \code{data.frame}. If package is not installed,
install with \code{install.packages(dependencies = TRUE)}.
\enumerate{
\item Check if the package is installed and load it
\item Retrieve the objects in the package
\item Use \code{purrr::map_lgl()} to apply \code{is.data.frame()} to each object in the
package. \code{map_lgl()} returns a logical vector with the same length as
the retrieved package objects.
}
}
\examples{
check_data_frame_in_pkg("dplyr")
check_data_frame_in_pkg("stringr")
}
\seealso{
\code{\link[=check_pkg_inst]{check_pkg_inst()}}
}
