% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inst_pkg.R
\name{check_inst_pkg}
\alias{check_inst_pkg}
\title{Check if package is loaded, if not load it}
\usage{
check_inst_pkg(pkg, quiet = FALSE)
}
\arguments{
\item{pkg}{name of package (a character vector)}

\item{quiet}{logical, print messages}
}
\value{
Package: \code{'name'} loaded or Loading package: \code{'name'}
}
\description{
Check if \code{pkg} is installed with \code{isNamespaceLoaded()}. If not, package is
installed with \code{requireNamespace(quietly = TRUE)}.
}
\examples{
# remove from ns
unloadNamespace("fs")
unloadNamespace("box")
# with single pkg
check_inst_pkg("fs")
# remove again
unloadNamespace("fs")
# with multiple pkgs
pkgs <- c("fs", "box")
check_inst_pkg(pkgs)
}
