% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_inst.R
\name{check_pkg_inst}
\alias{check_pkg_inst}
\title{Check if package is installed, if not install it}
\usage{
check_pkg_inst(pkg)
}
\arguments{
\item{pkg}{name of package (a character vector)}
}
\description{
Check if \code{pkg} is installed with \code{require()}. If not, package is
installed with \code{install.packages(quietly = TRUE)}.
}
\examples{
# remove.packages('janitor')
check_pkg_inst("janitor")
}
