% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_type_count.R
\name{col_type_count}
\alias{col_type_count}
\title{Get number of data.frame columns by type}
\usage{
col_type_count(df, type)
}
\arguments{
\item{df}{\code{data.frame} or \code{tibble}}

\item{type}{column type, one of: "num", "chr", "fct", "log",
"lst".}
}
\value{
count of columns by type
}
\description{
Get number of data.frame columns by type
}
\examples{
col_type_count(mtcars, "num")
}
