% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbap-package.R
\docType{package}
\name{dbap-package}
\alias{dbap}
\alias{dbap-package}
\title{dbap: Test Your Shiny App Package with RStudio'd debugger}
\description{
This shiny app-package was developed to demonstrate testing with testthat, shiny's testServer(), and the debugger.
}
\keyword{internal}
