% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_skim}
\alias{df_skim}
\title{Custom \code{skimr::skim()} for numeric variables}
\usage{
df_skim(df)
}
\description{
The custom \code{skimr::skim_with()}
}
\examples{
require(dplyr)
df_skim(dplyr::starwars)
}
