% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_select_vars.R
\name{filter_vars_fun}
\alias{filter_vars_fun}
\title{Filter variables by function}
\usage{
filter_vars_fun(data, filter)
}
\arguments{
\item{data}{\code{data.frame}/\code{tibble}}

\item{filter}{any \code{is.[type]} function (i.e., \code{is.numeric},
\code{is.character}, \code{is.logical}, etc.)}
}
\value{
data.frame/tibble filtered by function
}
\description{
Filter variables by function
}
