% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get dataset from package}
\usage{
get_data(ds, pkg)
}
\arguments{
\item{ds}{name of dataset}

\item{pkg}{name of package}
}
\value{
data object
}
\description{
Get dataset from package
}
\examples{
get_data("gss_cat", "forcats")
get_data("starwars", "dplyr")
}
