% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_results.R
\name{get_data_results}
\alias{get_data_results}
\title{Re-worked version of datasets() from (vcdExtra)}
\usage{
get_data_results(
  pkg,
  allClass = FALSE,
  incPackage = length(pkg) > 1,
  maxTitle = NULL
)
}
\arguments{
\item{pkg}{name of package}

\item{allClass}{logical, include all classes of the item (\code{TRUE}) or just
the last class (\code{FALSE})}

\item{incPackage}{logical, include package name in result?}

\item{maxTitle}{maximum length of data set Title}
}
\value{
metadata table of data object
}
\description{
Re-worked version of datasets() from (vcdExtra)
}
\examples{
get_data_results("lubridate", incPackage = TRUE)
get_data_results("dplyr", incPackage = TRUE)
}
