% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_col_count.R
\name{get_df_col_count}
\alias{get_df_col_count}
\title{Get number of data.frame columns by type}
\usage{
get_df_col_count(df, type)
}
\arguments{
\item{df}{\code{data.frame} or \code{tibble}}

\item{type}{column type, one of: "num", "chr", "fct", "log",
"lst".}
}
\value{
count of columns by type
}
\description{
Get number of data.frame columns by type
}
\examples{
get_df_col_count(mtcars, "num")
}
