% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ds_metadata.R
\name{get_ds_metadata}
\alias{get_ds_metadata}
\title{Get dataset metadata from packages}
\usage{
get_ds_metadata(pak)
}
\arguments{
\item{pak}{name of package(s)}
}
\value{
metadata table of datasets in package(s)
}
\description{
Get dataset metadata from packages
}
\examples{
require(lubridate)
require(forcats)
require(tidyr)
pkgs <- c("lubridate", "forcats")
get_ds_metadata(pak = pkgs)
get_ds_metadata("tidyr")
}
