% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_data.R
\name{get_pkg_df_names}
\alias{get_pkg_df_names}
\title{Get names of data.frames from package}
\usage{
get_pkg_df_names(pkg)
}
\arguments{
\item{pkg}{}
}
\value{
named vector of data.frames in package
}
\description{
Get names of data.frames from package
}
\examples{
get_pkg_df_names(pkg = "base")
get_pkg_df_names(pkg = "datasets")
}
