% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_data.R
\name{get_search_list_pkgs}
\alias{get_search_list_pkgs}
\title{Get all packages on search list}
\usage{
get_search_list_pkgs()
}
\value{
All items from \code{search()} with a \verb{package:} prefix
}
\description{
This function is meant to be used in combination with \code{check_df_in_pkg()}
}
\examples{
get_search_list_pkgs()

}
\seealso{
\code{\link[=check_df_in_pkg]{check_df_in_pkg()}}
}
