% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter.R
\name{gg_base}
\alias{gg_base}
\title{Base plot (ggplot2)}
\usage{
gg_base(df, x_var, y_var)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable}

\item{y_var}{y variable}
}
\value{
plot object
}
\description{
Base plot (ggplot2)
}
\examples{
require(dplyr)
starwars <- dplyr::starwars
gg_base(df = starwars, x_var = "height", y_var = "mass")
}
