% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter.R
\name{gg_color_scatter_facet}
\alias{gg_color_scatter_facet}
\title{Colored point plot (scatter-plot) with facets}
\usage{
gg_color_scatter_facet(df, x_var, y_var, col_var = NULL, facet_var = NULL, ...)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable (supplied to \code{ggplot2::aes(x = )})}

\item{y_var}{y variable (supplied to \code{ggplot2::aes(y = )})}

\item{col_var}{color variable (supplied to \code{ggplot2::geom_point(ggplot2::aes(color = ))})}

\item{facet_var}{facet variable (supplied to \code{ggplot2::geom_point(ggplot2::aes(color = ))})}

\item{...}{other arguments passed to (\code{ggplot2::facet_wrap(vars())})}
}
\value{
A \code{ggplot2} plot object
}
\description{
Colored point plot (scatter-plot) with facets
}
\examples{
require(dplyr)
gg_color_scatter_facet(
  df = dplyr::starwars,
  x_var = "height",
  y_var = "mass",
  col_var = "hair_color",
  facet_var = "gender",
  alpha = 1 / 3,
  size = 2
)
# compare with
ggplot2::ggplot(
  data = dplyr::starwars,
  mapping = ggplot2::aes(x = height, y = mass)
) +
  ggplot2::geom_point(aes(color = hair_color, group = gender),
    size = 2, alpha = 1 / 3
  ) +
  ggplot2::facet_wrap(. ~ gender) +
  ggplot2::theme_minimal() +
  ggplot2::theme(legend.position = "bottom")
gg_color_scatter_facet(
  df = dplyr::starwars,
  x_var = "height",
  y_var = "mass",
  col_var = "hair_color",
  facet_var = NULL,
  alpha = 1 / 3,
  size = 2
)
# compare with
ggplot2::ggplot(
  data = dplyr::starwars,
  mapping = ggplot2::aes(x = height, y = mass)
) +
  ggplot2::geom_point(aes(color = hair_color, group = gender),
    size = 2, alpha = 1 / 3
  ) +
  ggplot2::theme_minimal() +
  ggplot2::theme(legend.position = "bottom")
gg_color_scatter_facet(
  df = dplyr::starwars,
  x_var = "height",
  y_var = "mass",
  col_var = NULL,
  facet_var = NULL,
  alpha = 1 / 3,
  size = 2
)
# compare with
ggplot(
  data = dplyr::starwars,
  mapping = ggplot2::aes(x = height, y = mass)
) +
  ggplot2::geom_point(size = 2, alpha = 1 / 3) +
  ggplot2::theme_minimal() +
  ggplot2::theme(legend.position = "bottom")
}
