% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter.R
\name{gg_scatter}
\alias{gg_scatter}
\title{Point plot (scatter-plot)}
\usage{
gg_scatter(df, x_var, y_var, ...)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable (supplied to \code{ggplot2::aes(x = )})}

\item{y_var}{y variable (supplied to \code{ggplot2::aes(y = )})}

\item{...}{other arguments passed to \code{ggplot2::geom_point()}, outside of \code{ggplot2::aes()}}
}
\value{
A \code{ggplot2} plot object
}
\description{
Point plot (scatter-plot)
}
\examples{
require(palmerpenguins)
penguins <- palmerpenguins::penguins
gg_scatter(
  df = penguins,
  x_var = "bill_length_mm",
  y_var = "body_mass_g",
  alpha = 1 / 3,
  color = "#000000",
  size = 2
)
}
