% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_x_y_col_facet_title.R
\name{gg_x_y_col_facet_title}
\alias{gg_x_y_col_facet_title}
\title{Make x, y, color plot title}
\usage{
gg_x_y_col_facet_title(x, y, color, facets)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{color}{color variable}
}
\value{
String for plot title
}
\description{
Make x, y, color plot title
}
\examples{
gg_x_y_col_facet_title(
  x = "height",
  y = "mass",
  color = "hair_color"
)
}
