% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_x_y_color_title.R
\name{gg_x_y_color_title}
\alias{gg_x_y_color_title}
\title{Make x, y, color plot title}
\usage{
gg_x_y_color_title(x, y, color)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{color}{color variable}
}
\value{
String for plot title
}
\description{
Make x, y, color plot title
}
\examples{
require(ggplot2)
diamonds <- ggplot2::diamonds
mini_dmnds <- diamonds[sample(nrow(diamonds), size = 1000), ]
base <- ggplot2::ggplot(data = mini_dmnds,
  mapping = ggplot2::aes(x = carat, y = price, color = cut)) +
   ggplot2::geom_point()
base +
 ggplot2::labs(title = gg_x_y_color_title(
      x = "carat",
      y = "price",
      color = "cut"))
}
