% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_x_y_title.R
\name{gg_x_y_title}
\alias{gg_x_y_title}
\title{ggplot2 x, y plot title}
\usage{
gg_x_y_title(x, y)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{color}{color variable}
}
\value{
String for plot title
}
\description{
ggplot2 x, y plot title
}
\examples{
require(ggplot2)
diamonds <- ggplot2::diamonds
mini_dmnds <- diamonds[sample(nrow(diamonds), size = 1000), ]
base <- ggplot2::ggplot(data = mini_dmnds,
  mapping = ggplot2::aes(x = carat, y = price)) +
   ggplot2::geom_point()
base +
 ggplot2::labs(title =
    gg_x_y_title(
      x = "height",
      y = "mass"))
}
