% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter.R
\name{make_x_y_color_title}
\alias{make_x_y_color_title}
\title{Make x, y, color plot title}
\usage{
make_x_y_color_title(x, y, color)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{color}{color variable}
}
\value{
String for plot title
}
\description{
Make x, y, color plot title
}
\examples{
make_x_y_color_title(
  x = "imdb_rating",
  y = "audience_score",
  color = "mpaa_rating"
)
}
