% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter.R
\name{make_x_y_title}
\alias{make_x_y_title}
\title{Make x,y plot title}
\usage{
make_x_y_title(x, y)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{color}{color variable}
}
\value{
String for plot title
}
\description{
Make x,y plot title
}
\examples{
make_x_y_title(
  x = "imdb_rating",
  y = "audience_score"
)
}
