% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pkg_data.R
\name{mod_pkg_data_server}
\alias{mod_pkg_data_server}
\title{Package dataset server module}
\usage{
mod_pkg_data_server(id)
}
\arguments{
\item{id}{namespaced module id}
}
\value{
shiny module server function
}
\description{
Package dataset server module
}
