% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pkg_data_str.R
\name{mod_pkg_data_str_server}
\alias{mod_pkg_data_str_server}
\title{Select variable server module}
\usage{
mod_pkg_data_str_server(id, pkg_data)
}
\arguments{
\item{id}{namespaced module id}
}
\value{
shiny server module
}
\description{
Select variable server module
}
