% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pkg_data_str.R
\name{mod_pkg_data_str_ui}
\alias{mod_pkg_data_str_ui}
\title{Package data structure UI module}
\usage{
mod_pkg_data_str_ui(id)
}
\arguments{
\item{id}{namespaced module id}
}
\value{
shiny UI module
}
\description{
Package data structure UI module
}
