% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pkg_data.R
\name{mod_pkg_data_ui}
\alias{mod_pkg_data_ui}
\title{Package dataset UI}
\usage{
mod_pkg_data_ui(id)
}
\arguments{
\item{id}{namespaced module id}
}
\value{
shiny UI module
}
\description{
Package dataset UI
}
