% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_vars.R
\name{mod_select_vars_server}
\alias{mod_select_vars_server}
\title{Select variable server module}
\usage{
mod_select_vars_server(id, pkg_data)
}
\arguments{
\item{id}{namespaced module id}
}
\value{
shiny server module
}
\description{
Select variable server module
}
