% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_vars.R
\name{mod_select_vars_ui}
\alias{mod_select_vars_ui}
\title{Select variables UI module}
\usage{
mod_select_vars_ui(id)
}
\arguments{
\item{id}{namespaced module id}
}
\value{
shiny UI module
}
\description{
Select variables UI module
}
