% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_frame_nms.R
\name{pkg_data_frame_nms}
\alias{pkg_data_frame_nms}
\title{Get names of data.frames from package}
\usage{
pkg_data_frame_nms(pkg)
}
\arguments{
\item{pkg}{name of package}
}
\value{
named vector of data.frames in package
}
\description{
Get names of data.frames from package
}
\examples{
pkg_data_frame_nms(pkg = "base")
pkg_data_frame_nms(pkg = "datasets")
pkg_data_frame_nms(pkg = "lubridate")
}
