% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_object.R
\name{pkg_data_object}
\alias{pkg_data_object}
\title{Get dataset from package}
\usage{
pkg_data_object(ds, pkg)
}
\arguments{
\item{ds}{name of dataset}

\item{pkg}{name of package}
}
\value{
data object
}
\description{
Get dataset from package
}
\examples{
pkg_data_object("gss_cat", "forcats")
pkg_data_object("starwars", "dplyr")
}
