% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_results.R
\name{pkg_data_results}
\alias{pkg_data_results}
\title{Get results from \verb{data(package = [package])}}
\usage{
pkg_data_results(pkg)
}
\arguments{
\item{pkg}{name of package}
}
\value{
metadata table of data object
}
\description{
Get results from \verb{data(package = [package])}
}
\examples{
pkg_data_results("lubridate")
pkg_data_results("dplyr")
}
