% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_str.R
\name{pkg_data_str}
\alias{pkg_data_str}
\title{Get dataset structures from R package}
\usage{
pkg_data_str(pkg)
}
\arguments{
\item{pkg}{name of package(s)}
}
\value{
dataset structure \code{tibble}
}
\description{
Get dataset structures from R package
}
\section{The returned object has the following columns}{

\itemize{
\item \code{Package}: name of package
\item \code{Dataset}: name of dataset
\item \code{Class}: class of dataset
\item \code{Columns}: total columns
\item \code{Rows}: total rows
\item \code{Logical}: total logical columns
\item \code{Numeric}: total numeric columns
\item \code{Character}: total character columns
\item \code{Factor}: total factor columns
\item \code{List}: total list columns
}
}

\examples{
require(forcats)
pkg_data_str("forcats")
require(dplyr)
pkg_data_str(c("dplyr", "forcats"))
}
