% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ds_str.R, R/pkg_data_strs.R
\name{pkg_data_strs}
\alias{pkg_data_strs}
\title{Get dataset structures from R package}
\usage{
pkg_data_strs(pkg)

pkg_data_strs(pkg)
}
\arguments{
\item{pkg}{name of package}
}
\value{
dataset structure \code{tibble}

data object
}
\description{
Get dataset structures from R package

Get dataset structure from package
}
\section{The returned object has the following columns}{

\itemize{
\item \code{Package}: name of package
\item \code{Dataset}: name of dataset
\item \code{Class}: class of dataset
\item \code{Columns}: total columns
\item \code{Rows}: total rows
\item \code{Logical}: total logical columns
\item \code{Numeric}: total numeric columns
\item \code{Character}: total character columns
\item \code{Factor}: total factor columns
\item \code{List}: total list columns
}
}

\examples{
require(forcats)
pkg_data_strs("forcats")
require(dplyr)
pkg_data_strs(c("dplyr", "forcats"))
require(forcats)
require(dplyr)
pkg_data_strs("forcats")
pkg_data_strs(c("dplyr", "forcats"))
}
