% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs_nms_with_data_frames.R
\name{pkg_nms_with_data_frames}
\alias{pkg_nms_with_data_frames}
\title{Get packages with data.frames}
\usage{
pkg_nms_with_data_frames()
}
\value{
A named vector of packages with \code{data.frame} objects
}
\description{
Get packages with data.frames
}
\examples{
pkg_nms_with_data_frames()
}
