% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_list_pkg_nms.R
\name{search_list_pkg_nms}
\alias{search_list_pkg_nms}
\title{Get all packages on search list}
\usage{
search_list_pkg_nms()
}
\value{
All items from \code{search()} with a \verb{package:} prefix
}
\description{
This function is meant to be used in combination with \code{check_data_frame_in_pkg()}
}
\examples{
search_list_pkg_nms()

}
\seealso{
\code{\link[=check_data_frame_in_pkg]{check_data_frame_in_pkg()}}
}
