% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg_data.R
\name{check_inst_pkg}
\alias{check_inst_pkg}
\title{Check if package is loaded, if not load it}
\usage{
check_inst_pkg(pkg)
}
\arguments{
\item{pkg}{name of package (a character vector)}
}
\value{
Package: \code{'name'} loaded or Loading package: \code{'name'}
}
\description{
Check if \code{pkg} is installed. If not, package is installed with
\code{install.packages(dependencies = TRUE)} and loaded with
\code{library(pkg, character.only = TRUE)}
}
\examples{
check_inst_pkg("plotly")
}
