% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ds_metadata.R
\name{get_data}
\alias{get_data}
\title{Get dataset from package}
\usage{
get_data(x, pkg)
}
\arguments{
\item{x}{name of dataset}

\item{pkg}{name of package}
}
\value{
data object
}
\description{
Get dataset from package
}
\examples{
get_data("lakers", "lubridate")
get_data("starwars", "dplyr")
}
