% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ds_metadata.R
\name{get_data_results}
\alias{get_data_results}
\title{Get results from \verb{data(package = [package])}}
\usage{
get_data_results(pkg)
}
\arguments{
\item{pkg}{name of package}
}
\value{
metadata table of data object
}
\description{
Get results from \verb{data(package = [package])}
}
\examples{
get_data_results("lubridate")
get_data_results("dplyr")
}
