% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_base.R
\name{gg_base}
\alias{gg_base}
\title{Base plot (ggplot2)}
\usage{
gg_base(df, x_var, y_var)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable}

\item{y_var}{y variable}
}
\value{
plot object
}
\description{
Base plot (ggplot2)
}
\examples{
require(ggplot2)
diamonds <- ggplot2::diamonds
mini_dmnds <- diamonds[sample(nrow(diamonds), size = 1000), ]
gg_base(df = mini_dmnds, x_var = "carat", y_var = "price")
}
