% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter.R
\name{gg_color_scatter}
\alias{gg_color_scatter}
\title{Colored point plot (scatter-plot)}
\usage{
gg_color_scatter(df, x_var, y_var, col_var, ...)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable (supplied to \code{ggplot2::aes(x = )})}

\item{y_var}{y variable (supplied to \code{ggplot2::aes(y = )})}

\item{col_var}{color variable (supplied to \code{ggplot2::geom_point(ggplot2::aes(color = ))})}

\item{...}{other arguments passed to \code{ggplot2::geom_point()}, outside of \code{ggplot2::aes()}}
}
\value{
A \code{ggplot2} plot object
}
\description{
Colored point plot (scatter-plot)
}
\examples{
require(palmerpenguins)
penguins <- palmerpenguins::penguins
gg_color_scatter(
  df = penguins,
  x_var = "bill_length_mm",
  y_var = "body_mass_g",
  col_var = "species",
  alpha = 1 / 3,
  size = 2
)
}
