% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ds_str.R, R/pkg_data_strs.R
\name{pkg_data_strs}
\alias{pkg_data_strs}
\title{Get dataset structures from R package}
\usage{
pkg_data_strs(pkg)

pkg_data_strs(pkg)
}
\arguments{
\item{pkg}{name of package}
}
\value{
\code{tibble} with the following columns:

\describe{
\item{Package}{name of package}
\item{Dataset}{name of dataset}
\item{Class}{class of dataset}
\item{Columns}{total columns}
\item{Rows}{total rows}
\item{Logical}{total logical columns}
\item{Numeric}{total numeric columns}
\item{Character}{total character columns}
\item{Factor}{total factor columns}
\item{List}{total list columns}
}

data object
}
\description{
Get dataset structures from R package

Get dataset structure from package
}
\examples{
require(forcats)
pkg_data_strs("forcats")
require(dplyr)
pkg_data_strs(c("dplyr", "forcats"))
require(forcats)
require(dplyr)
pkg_data_strs("forcats")
pkg_data_strs(c("dplyr", "forcats"))
}
